+++ src/tomcatmanager/tomcat_manager.py
@@ -148,33 +148,16 @@ class TomcatManager:
         # ask nicely for people not to access the password attribute
         self._password = None
 
-        # same with the certificates
+        # same with the certificates for client side authentication
         self._cert = None
+
+        # and the CA bundle to verify server SSL/TLS certifications,
+        # or False to skip verification
+        self._verify = None
 
         # where we keep track of the version of tomcat we are connected to
         # this is set by connect()
         self._tomcat_major_minor = None
-
-        self.verify = True
-        """Verify server SSL/TLS certificates using the given CA directory or file.
-
-        If the URL uses the ``https`` protocol, the default behavior is to validate
-        the SSL/TLS certificate chain.
-
-        To validate with your own certificate authority bundle, set this to the
-        path to a certificate authority bundle file or a directory of certificates
-        of trusted certificate authorities.
-
-        For testing purposes server certificate validation can occasionally be
-        undesirable. You can turn off the validation by setting this to ``False``.
-        If you find yourself in this situation, rather than disabling certificate
-        verification, I suggest you see if the free certificates issued by
-        `Let's Encrypt <https://letsencrypt>`_ can meet your needs.
-
-        See `SSL Certificate Verification
-        <https://docs.python-requests.org/en/master/user/advanced/#ssl-cert-verification>`_
-        in the Requests documentation for more information.
-        """
 
         self.timeout = 10.0
         """Seconds to wait before giving up on network operations. Can be a
@@ -192,28 +175,43 @@ class TomcatManager:
     def url(self) -> str:
         """Url of the Tomcat Manager web application we are connected to.
 
-        This attribute is set by the :meth:`.connect` method. Look there for more info.
+        This attribute is set by the :meth:`.connect` method. Look there for
+        more info.
         """
         return self._url
 
     @property
     def user(self) -> str:
-        """User we successfully authenticated to the Tomcat Manager web application
-        with.
+        """User we successfully authenticated to the Tomcat Manager web
+        application with.
 
-        This attribute is set by the :meth:`.connect` method. Look there for more info.
+        This attribute is set by the :meth:`.connect` method. Look there for
+        more info.
         """
         return self._user
 
     @property
     def cert(self) -> Union[str, Tuple[str, str]]:
-        """Path to file containing client certificate and private key to use for
-        SSL/TLS authentication. Alternatively, use a tuple containing the path to
-        the certificate and the path to the private key. See `Client Side Certificates
-        <https://docs.python-requests.org/en/master/user/advanced/#client-side-certificates>`_
-        in the Requests documentation for more information.
+        """Client side SSL/TLS certificates we authenticated to the Tomcat Manager
+        web application with.
+
+        This attribute is set by the :meth:`.connect` method. Look there for
+        more info.
+
+        .. versionadded:: 2.1.0
         """
         return self._cert
+
+    @property
+    def verify(self) -> Union[str, bool]:
+        """The certificate authority directory or bundle to use to verify
+        server SSL/TLS certificates. If ``False`` no verification is performed.
+
+        This attribute is set by the :meth:`.connect` method. Look there for more info.
+
+        .. versionadded:: 2.1.0
+        """
+        return self._verify
 
     @property
     def tomcat_major_minor(self) -> TomcatMajorMinor:
@@ -307,17 +305,19 @@ class TomcatManager:
         url: str,
         user: str = "",
         password: str = "",
-        timeout: float = None,
         *,
         cert: Union[str, Tuple[str, str]] = None,
+        verify: Union[str, bool] = True,
+        timeout: float = None,
     ) -> TomcatManagerResponse:
         """Connect to the manager application running in a Tomcat server.
 
         :param url:      url where the Tomcat Manager web application is deployed
         :param user:     (optional) user to authenticate with
         :param password: (optional) password to authenticate with
-        :param timeout:  timeout in seconds for network operations
         :param cert:     client side certificates to use for SSL/TLS authentication
+        :param verify:   verify server SSL/TLS certificates
+        :param timeout:  timeout in seconds for network operations
         :return:         :class:`~tomcatmanager.models.TomcatManagerResponse`
                          object with an additional ``server_info`` attribute
 
@@ -335,7 +335,7 @@ class TomcatManager:
           you are connected to
         - allows you to inspect the response so you can see why you can't connect
 
-        Usage:
+        **Usage**
 
         >>> import tomcatmanager as tm
         >>> url = 'http://localhost:808099/manager'
@@ -384,14 +384,22 @@ class TomcatManager:
         is no persistent connection of any kind, there is no disconnect method and
         no cleanup to perform when you are done using a server.
 
+        If you discard or don't save the return object from this method, you can call
+        :meth:`is_connected` to check if you are connected.
+
+        **Authentication**
+
         The only way to validate the URL and authentication credentials is to
         make an HTTP request to the server and see if it returns successfully.
         Internally this method tries to retrieve ``/manager/text/serverinfo``.
 
+        Typically authentication is done via user and password, pass those
+        parameters to utilize HTTP Basic authentication.
+
         To authenticate with a SSL/TLS server using a client certificate and key, pass
         the path to a single file containing the private key and certificate in the
         ``cert`` parameter. As an alternative, you can pass a tuple of both files'
-        paths. If you use this, you shouldn't also pass a username and password.
+        paths. If you use this, you should probably use a null username and password.
         See `Client Side Certificates
         <https://docs.python-requests.org/en/master/user/advanced/#client-side-certificates>`_
         in the Requests documentation for additional details.
@@ -402,8 +410,27 @@ class TomcatManager:
             Requests library used for network communication does not support using
             encrypted keys.
 
-        Passing a timeout parameter to this method has the side effect of setting the
-        :attr:`timeout` attribute on this object.
+        If the URL uses the ``https`` protocol, the default behavior is to validate
+        the server SSL/TLS certificate chain.
+
+        To validate with your own certificate authority bundle, set the
+        ``verify`` parameter to the path to a certificate authority bundle file
+        or a directory of certificates of trusted certificate authorities.
+
+        For testing purposes server certificate validation can occasionally be
+        undesirable. You can turn off the validation by setting ``verify`` to
+        ``False``. If you find yourself in this situation, rather than disabling
+        certificate verification, I suggest you see if the free certificates issued
+        by `Let's Encrypt <https://letsencrypt>`_ can meet your needs.
+
+        See `SSL Certificate Verification
+        <https://docs.python-requests.org/en/master/user/advanced/#ssl-cert-verification>`_
+        in the Requests documentation for more information.
+
+        **Side Effects**
+
+        Passing a ``timeout`` parameter to this method has the side effect of
+        setting the :attr:`timeout` attribute on this object.
 
         Requesting url's via http can also result in redirection to another url. If
         that occurs, the new url, not the one you passed, will be stored in the
@@ -412,16 +439,15 @@ class TomcatManager:
         If you pass authentication credentials and the connection is successful, the
         user will be stored in the :attr:`user` attribute.
 
+        If you pass an authentication key and certificate in the ``cert`` parameter
+        and the connection is successful, the information will be stored in the
+        :attr:`cert` attribute.
+
         Upon successful connection, an instance of :class:`.TomcatMajorMinor` will be
         stored in :attr:`tomcat_major_minor` indicating the major version of Tomcat
         running on the server. Further details about the server are available in the
         `server_info` attribute of the returned response.
 
-        If you discard or don't save the return object from this method, you can call
-        :meth:`is_connected` to check if you are connected.
-
-        If you want to raise more exceptions see
-        :meth:`.TomcatManagerResponse.raise_for_status`.
         """
         if timeout:
             self.timeout = timeout
@@ -431,17 +457,27 @@ class TomcatManager:
         self._user = user
         self._password = password
         self._cert = cert
+        self._verify = verify
         # don't use server_info() because it will fail because it won't
         # think we are connected to the server yet
-        r = self._get("serverinfo")
-        r.server_info = ServerInfo(result=r.result)
-
-        if r.ok:
-            self._set_server_attrs(r)
-        else:
-            # don't save the parameters if we don't succeed
+        try:
+            r = self._get("serverinfo")
+            r.server_info = ServerInfo(result=r.result)
+            if r.ok:
+                self._tomcat_major_minor = r.server_info.tomcat_major_minor
+                # _get added /text/serverinfo onto the end of the passed in url
+                # we may have been redirected, and we want to store the new
+                # url, not the one passed in
+                match = re.search(r"(.*)/text/serverinfo$", r.response.url)
+                if match:
+                    self._url = match.group(1)
+            else:
+                # don't save the parameters if we don't succeed
+                self._clear_server_attrs()
+            return r
+        except Exception as exc:
             self._clear_server_attrs()
-        return r
+            raise exc
 
     ###
     # managing applications
@@ -919,7 +955,7 @@ class TomcatManager:
             params={"XML": "true"},
             timeout=self.timeout,
             verify=self.verify,
-            cert=self._cert,
+            cert=self.cert,
         )
         r.result = r.response.text
         r.status_xml = r.result
@@ -1074,21 +1110,9 @@ class TomcatManager:
         self._user = None
         self._password = None
         self._url = None
+        self._cert = None
+        self._verify = None
         self._tomcat_major_minor = None
-
-    def _set_server_attrs(self, response: TomcatManagerResponse):
-        """
-        Set private attributes based on whether we connected to the url.
-
-        Intended to be called from connect() and is_connected()
-        """
-        self._tomcat_major_minor = response.server_info.tomcat_major_minor
-        # _get added /text/serverinfo onto the end of the passed in url
-        # we may have been redirected, and we want to store the new
-        # url, not the one passed in
-        match = re.search(r"(.*)/text/serverinfo$", response.response.url)
-        if match:
-            self._url = match.group(1)
 
     def _get(self, cmd: str, payload: dict = None) -> TomcatManagerResponse:
         """
@@ -1106,13 +1130,17 @@ class TomcatManager:
             url = base + "/text/" + cmd
         else:
             url = ""
+        authinfo = None
+        if self._user and self._password:
+            authinfo = (self._user, self._password)
+
         r = TomcatManagerResponse()
         r.response = requests.get(
             url,
-            auth=(self._user, self._password),
+            auth=authinfo,
             params=payload,
             timeout=self.timeout,
             verify=self.verify,
-            cert=self._cert,
+            cert=self.cert,
         )
         return r